requirements.txt
streamlit
pillow
pytesseract
opencv-python
numpy
pandas
APScheduler
python-dotenv
scikit-learn
joblib
dateparser
